[engraving] moved layout Beam
