+++ src/engraving/CMakeLists.txt
@@ -156,8 +156,10 @@ set(MODULE_SRC
     ${CMAKE_CURRENT_LIST_DIR}/layout/layoutlyrics.h
     ${CMAKE_CURRENT_LIST_DIR}/layout/layoutmeasure.cpp
     ${CMAKE_CURRENT_LIST_DIR}/layout/layoutmeasure.h
-    ${CMAKE_CURRENT_LIST_DIR}/layout/layoutbeams.cpp
-    ${CMAKE_CURRENT_LIST_DIR}/layout/layoutbeams.h
+    ${CMAKE_CURRENT_LIST_DIR}/layout/beamlayout.cpp
+    ${CMAKE_CURRENT_LIST_DIR}/layout/beamlayout.h
+    ${CMAKE_CURRENT_LIST_DIR}/layout/beamtremololayout.cpp
+    ${CMAKE_CURRENT_LIST_DIR}/layout/beamtremololayout.h
     ${CMAKE_CURRENT_LIST_DIR}/layout/layouttuplets.cpp
     ${CMAKE_CURRENT_LIST_DIR}/layout/layouttuplets.h
     ${CMAKE_CURRENT_LIST_DIR}/layout/layoutchords.cpp
