+++ src/engraving/layout/layoutpage.cpp
@@ -47,7 +47,7 @@
 #include "libmscore/tuplet.h"
 
 #include "layoutsystem.h"
-#include "layoutbeams.h"
+#include "beamlayout.h"
 #include "layouttuplets.h"
 #include "verticalgapdata.h"
 
@@ -273,7 +273,7 @@ void LayoutPage::collectPage(const LayoutOptions& options, LayoutContext& ctx)
                             continue;
                         }
                         ChordRest* cr = toChordRest(e);
-                        if (LayoutBeams::notTopBeam(cr)) {                           // layout cross staff beams
+                        if (BeamLayout::notTopBeam(cr)) {                           // layout cross staff beams
                             cr->beam()->layout();
                         }
                         if (LayoutTuplets::notTopTuplet(cr)) {
