+++ src/engraving/layout/tlayout.h
@@ -33,6 +33,7 @@ class Articulation;
 
 class BagpipeEmbellishment;
 class BarLine;
+class Beam;
 
 class TLayout
 {
@@ -47,6 +48,8 @@ public:
     static void layout(BagpipeEmbellishment* item, LayoutContext& ctx);
     static void layout(BarLine* item, LayoutContext& ctx);
     static void layout2(BarLine* item, LayoutContext& ctx);
+    static void layout(Beam* item, LayoutContext& ctx);
+    static void layout1(Beam* item, LayoutContext& ctx);
 
 private:
     static void layoutSingleGlyphAccidental(Accidental* item, LayoutContext& ctx);
