+++ src/engraving/libmscore/beam.cpp
@@ -30,6 +30,7 @@
 #include "realfn.h"
 
 #include "draw/types/brush.h"
+#include "layout/tlayout.h"
 
 #include "actionicon.h"
 #include "chord.h"
@@ -49,7 +50,7 @@
 #include "tremolo.h"
 #include "tuplet.h"
 
-#include "layout/layoutbeams.h"
+#include "layout/beamlayout.h"
 
 #include "log.h"
 
@@ -271,226 +272,14 @@ void Beam::move(const PointF& offset)
 }
 
 //---------------------------------------------------------
-//   layout1
-//    - remove beam segments
-//    - detach from system
-//    - calculate stem direction and set chord
-//---------------------------------------------------------
-
-void Beam::layout1()
-{
-    resetExplicitParent();  // parent is System
-
-    const StaffType* staffType = this->staffType();
-    _tab = (staffType && staffType->isTabStaff()) ? staffType : nullptr;
-    _isBesideTabStaff = _tab && !_tab->stemless() && !_tab->stemThrough();
-
-    // TAB's with stem beside staves have special layout
-    if (_isBesideTabStaff) {
-        _up = !_tab->stemsDown();
-        _slope = 0.0;
-        _cross = false;
-        _minMove = 0;
-        _maxMove = 0;
-        for (ChordRest* cr : _elements) {
-            if (cr->isChord()) {
-                _up = cr->up();
-                break;
-            }
-        }
-        return;
-    }
-
-    if (staff()->isDrumStaff(Fraction(0, 1))) {
-        if (_direction != DirectionV::AUTO) {
-            _up = _direction == DirectionV::UP;
-        } else if (_isGrace) {
-            _up = true;
-        } else {
-            // search through the beam for the first chord with explicit stem direction and use that.
-            // if there is no explicit stem direction, default to the direction of the first stem.
-            bool firstUp = false;
-            bool firstChord = true;
-            for (ChordRest* cr :_elements) {
-                if (cr->isChord()) {
-                    DirectionV crDirection = toChord(cr)->stemDirection();
-                    if (crDirection != DirectionV::AUTO) {
-                        _up = crDirection == DirectionV::UP;
-                        break;
-                    } else if (firstChord) {
-                        firstUp = cr->up();
-                        firstChord = false;
-                    }
-                }
-                _up = firstUp;
-            }
-        }
-        for (ChordRest* cr : _elements) {
-            cr->computeUp();
-            if (cr->isChord()) {
-                toChord(cr)->layoutStem();
-            }
-        }
-        return;
-    }
-
-    _minMove = std::numeric_limits<int>::max();
-    _maxMove = std::numeric_limits<int>::min();
-    double mag = 0.0;
-
-    _notes.clear();
-    staff_idx_t staffIdx = mu::nidx;
-    for (ChordRest* cr : _elements) {
-        double m = cr->isSmall() ? score()->styleD(Sid::smallNoteMag) : 1.0;
-        mag = std::max(mag, m);
-        if (cr->isChord()) {
-            Chord* chord = toChord(cr);
-            staffIdx = chord->vStaffIdx();
-            int i = chord->staffMove();
-            _minMove = std::min(_minMove, i);
-            _maxMove = std::max(_maxMove, i);
-
-            for (int distance : chord->noteDistances()) {
-                _notes.push_back(distance);
-            }
-        }
-    }
-
-    std::sort(_notes.begin(), _notes.end());
-    setMag(mag);
-
-    //
-    // determine beam stem direction
-    //
-    if (_elements.empty()) {
-        return;
-    }
-    ChordRest* firstNote = _elements.front();
-    Measure* measure = firstNote->measure();
-    bool hasMultipleVoices = measure->hasVoices(firstNote->staffIdx(), tick(), ticks());
-    if (_direction != DirectionV::AUTO) {
-        _up = _direction == DirectionV::UP;
-    } else if (_maxMove > 0) {
-        _up = false;
-    } else if (_minMove < 0) {
-        _up = true;
-    } else if (_isGrace) {
-        if (hasMultipleVoices) {
-            _up = firstNote->track() % 2 == 0;
-        } else {
-            _up = true;
-        }
-    } else if (_notes.size()) {
-        if (hasMultipleVoices) {
-            _up = firstNote->track() % 2 == 0;
-        } else {
-            if (const Chord* chord = findChordWithCustomStemDirection()) {
-                _up = chord->stemDirection() == DirectionV::UP;
-            } else {
-                std::set<int> noteSet(_notes.begin(), _notes.end());
-                std::vector<int> notes(noteSet.begin(), noteSet.end());
-                _up = Chord::computeAutoStemDirection(notes) > 0;
-            }
-        }
-    } else {
-        _up = true;
-    }
-
-    int middleStaffLine = firstNote->staffType()->middleLine();
-    for (size_t i = 0; i < _notes.size(); i++) {
-        _notes[i] += middleStaffLine;
-    }
-
-    _cross = _minMove != _maxMove;
-    bool isEntirelyMoved = false;
-    if (_minMove == _maxMove && _minMove != 0) {
-        isEntirelyMoved = true;
-        setStaffIdx(staffIdx);
-        if (_direction == DirectionV::AUTO) {
-            _up = _maxMove > 0;
-        }
-    } else if (_elements.size()) {
-        setStaffIdx(_elements.at(0)->staffIdx());
-    }
-
-    _slope = 0.0;
-
-    for (ChordRest* cr : _elements) {
-        const bool staffMove = cr->isChord() ? toChord(cr)->staffMove() : false;
-        if (!_cross || !staffMove) {
-            if (cr->up() != _up) {
-                cr->setUp(isEntirelyMoved ? _up : (_up != staffMove));
-                if (cr->isChord()) {
-                    toChord(cr)->layoutStem();
-                }
-            }
-        }
-    }
-}
-
-//---------------------------------------------------------
 //   layout
 //   TODO - document what this function does
 //---------------------------------------------------------
 
 void Beam::layout()
 {
-    // all of the beam layout code depends on _elements being in order by tick
-    // this may not be the case if two cr's were recently swapped.
-    std::sort(_elements.begin(), _elements.end(),
-              [](const ChordRest* a, const ChordRest* b) -> bool {
-        return a->tick() < b->tick();
-    });
-    System* system = _elements.front()->measure()->system();
-    setParent(system);
-
-    std::vector<ChordRest*> crl;
-
-    size_t n = 0;
-    for (ChordRest* cr : _elements) {
-        auto newSystem = cr->measure()->system();
-        if (newSystem && newSystem != system) {
-            SpannerSegmentType st;
-            if (n == 0) {
-                st = SpannerSegmentType::BEGIN;
-            } else {
-                st = SpannerSegmentType::MIDDLE;
-            }
-            ++n;
-            if (fragments.size() < n) {
-                fragments.push_back(new BeamFragment);
-            }
-            layout2(crl, st, static_cast<int>(n) - 1);
-            crl.clear();
-            system = cr->measure()->system();
-        }
-        crl.push_back(cr);
-    }
-    setbbox(RectF());
-    if (!crl.empty()) {
-        SpannerSegmentType st;
-        if (n == 0) {
-            st = SpannerSegmentType::SINGLE;
-        } else {
-            st = SpannerSegmentType::END;
-        }
-        if (fragments.size() < (n + 1)) {
-            fragments.push_back(new BeamFragment);
-        }
-        layout2(crl, st, static_cast<int>(n));
-
-        double lw2 = _beamWidth / 2.0;
-
-        for (const BeamSegment* bs : _beamSegments) {
-            PolygonF a(4);
-            a[0] = PointF(bs->line.x1(), bs->line.y1());
-            a[1] = PointF(bs->line.x2(), bs->line.y2());
-            a[2] = PointF(bs->line.x2(), bs->line.y2());
-            a[3] = PointF(bs->line.x1(), bs->line.y1());
-            RectF r(a.boundingRect().adjusted(0.0, -lw2, 0.0, lw2));
-            addbbox(r);
-        }
-    }
+    LayoutContext ctx(score());
+    TLayout::layout(this, ctx);
 }
 
 //---------------------------------------------------------
@@ -557,279 +346,6 @@ void Beam::setTremAnchors()
     }
 }
 
-void Beam::createBeamSegment(ChordRest* startCr, ChordRest* endCr, int level)
-{
-    const bool isFirstSubgroup = startCr == _elements.front();
-    const bool isLastSubgroup = endCr == _elements.back();
-    const bool firstUp = startCr->up();
-    const bool lastUp = endCr->up();
-    bool overallUp = _up;
-    if (isFirstSubgroup == isLastSubgroup) {
-        // this subgroup is either the only one in the beam, or in the middle
-        if (firstUp == lastUp) {
-            // the "outside notes" of this subgroup go the same direction so use them
-            // to determine the side of the beams
-            overallUp = firstUp;
-        } else {
-            // no perfect way to solve this problem, for now we'll base it on the number of
-            // up and down stemmed notes in this subgroup
-            int upStems, downStems;
-            upStems = downStems = 0;
-            for (ChordRest* cr : _elements) {
-                if (!cr->isChord() || cr->tick() < startCr->tick()) {
-                    continue;
-                }
-                if (cr->tick() > endCr->tick()) {
-                    break;
-                }
-
-                ++(toChord(cr)->up() ? upStems : downStems);
-
-                if (cr == endCr) {
-                    break;
-                }
-            }
-            if (upStems == downStems) {
-                // we are officially bamboozled. for now we can just use the default
-                // direction based on the staff we're on I guess
-                overallUp = _up;
-            } else {
-                // use the direction with the most stems
-                overallUp = upStems > downStems;
-            }
-        }
-    } else if (isFirstSubgroup) {
-        overallUp = lastUp;
-    } else if (isLastSubgroup) {
-        overallUp = firstUp;
-    }
-
-    const double startX = _layoutInfo.chordBeamAnchorX(startCr, BeamTremoloLayout::ChordBeamAnchorType::Start);
-    const double endX = _layoutInfo.chordBeamAnchorX(endCr, BeamTremoloLayout::ChordBeamAnchorType::End);
-
-    double startY = _slope * (startX - _startAnchor.x()) + _startAnchor.y() - pagePos().y();
-    double endY = _slope * (endX - _startAnchor.x()) + _startAnchor.y() - pagePos().y();
-
-    int beamsBelow = 0; // how many beams below level 0?
-    int beamsAbove = 0; // how many beams above level 0?
-
-    // avoid adjusting for beams on opposite side of level 0
-    if (level != 0) {
-        for (const BeamSegment* beam : _beamSegments) {
-            if (beam->level == 0 || beam->endTick < startCr->tick() || beam->startTick > endCr->tick()) {
-                continue;
-            }
-            ++(beam->above ? beamsAbove : beamsBelow);
-        }
-
-        const int upValue = overallUp ? -1 : 1;
-        const int extraBeamAdjust = overallUp ? beamsAbove : beamsBelow;
-        const double verticalOffset = _beamDist * (level - extraBeamAdjust) * upValue;
-
-        if (RealIsEqual(_grow1, _grow2)) {
-            startY -= verticalOffset * _grow1;
-            endY -= verticalOffset * _grow1;
-        } else {
-            // Feathered beams
-            double startProportionAlongX = (startX - _startAnchor.x()) / (_endAnchor.x() - _startAnchor.x());
-            double endProportionAlongX = (endX - _startAnchor.x()) / (_endAnchor.x() - _startAnchor.x());
-
-            double grow1 = startProportionAlongX * (_grow2 - _grow1) + _grow1;
-            double grow2 = endProportionAlongX * (_grow2 - _grow1) + _grow1;
-
-            startY -= verticalOffset * grow1;
-            endY -= verticalOffset * grow2;
-        }
-    }
-
-    BeamSegment* b = new BeamSegment(this);
-    b->above = !overallUp;
-    b->level = level;
-    b->line = LineF(startX, startY, endX, endY);
-    b->startTick = startCr->tick();
-    b->endTick = endCr->tick();
-    _beamSegments.push_back(b);
-
-    if (level > 0) {
-        ++(b->above ? beamsAbove : beamsBelow);
-    }
-
-    // extend stems properly
-    for (ChordRest* cr : _elements) {
-        if (!cr->isChord() || cr->tick() < startCr->tick()) {
-            continue;
-        }
-        if (cr->tick() > endCr->tick()) {
-            break;
-        }
-
-        Chord* chord = toChord(cr);
-        double addition = 0.0;
-
-        if (level > 0) {
-            double grow = _grow1;
-            if (!RealIsEqual(_grow1, _grow2)) {
-                double anchorX = _layoutInfo.chordBeamAnchorX(chord, BeamTremoloLayout::ChordBeamAnchorType::Middle);
-                double proportionAlongX = (anchorX - _startAnchor.x()) / (_endAnchor.x() - _startAnchor.x());
-                grow = proportionAlongX * (_grow2 - _grow1) + _grow1;
-            }
-
-            int extraBeamAdjust = cr->up() ? beamsBelow : beamsAbove;
-            addition = grow * (level - extraBeamAdjust) * _beamDist;
-        }
-
-        if (level == 0 || !RealIsEqual(addition, 0.0)) {
-            _layoutInfo.extendStem(chord, addition);
-        }
-
-        if (chord == endCr) {
-            break;
-        }
-    }
-}
-
-bool Beam::calcIsBeamletBefore(Chord* chord, int i, int level, bool isAfter32Break, bool isAfter64Break) const
-{
-    // if first or last chord in beam group
-    if (i == 0) {
-        return false;
-    } else if (i == static_cast<int>(_elements.size()) - 1) {
-        return true;
-    }
-    // if first or last chord in tuplet
-    Tuplet* tuplet = chord->tuplet();
-    if (tuplet && chord == tuplet->elements().front()) {
-        return false;
-    } else if (tuplet && chord == tuplet->elements().back()) {
-        return true;
-    }
-
-    // next note has a beam break
-    ChordRest* nextChordRest = _elements[i + 1];
-    ChordRest* currChordRest = _elements[i];
-    ChordRest* prevChordRest = _elements[i - 1];
-    if (nextChordRest->isChord()) {
-        bool nextBreak32 = false;
-        bool nextBreak64 = false;
-        bool currBreak32 = false;
-        bool currBreak64 = false;
-        calcBeamBreaks(currChordRest, prevChordRest, prevChordRest->beams(), currBreak32, currBreak64);
-        calcBeamBreaks(nextChordRest, currChordRest, level, nextBreak32, nextBreak64);
-        if ((nextBreak32 && level >= 1) || (!currBreak32 && nextBreak64 && level >= 2)) {
-            return true;
-        }
-    }
-
-    // if previous or next chord has more beams, point in that direction
-    int previousChordLevel = -1;
-    int nextChordLevel = -1;
-    int previousOffset = 1;
-    while (i - previousOffset >= 0) {
-        ChordRest* previous = _elements[i - previousOffset];
-        if (previous->isChord()) {
-            previousChordLevel = toChord(previous)->beams();
-            if (isAfter32Break) {
-                previousChordLevel = std::min(previousChordLevel, 1);
-            } else if (isAfter64Break) {
-                previousChordLevel = std::min(previousChordLevel, 2);
-            }
-            break;
-        }
-        ++previousOffset;
-    }
-
-    int nextOffset = 1;
-    while (i + nextOffset < static_cast<int>(_elements.size())) {
-        ChordRest* next = _elements[i + nextOffset];
-        if (next->isChord()) {
-            nextChordLevel = toChord(next)->beams();
-            break;
-        }
-        ++nextOffset;
-    }
-    int chordLevelDifference = nextChordLevel - previousChordLevel;
-    if (chordLevelDifference != 0) {
-        return chordLevelDifference < 0;
-    }
-
-    // if the chord ends a subdivision of the beat
-    Fraction baseTick = tuplet ? tuplet->tick() : chord->measure()->tick();
-    Fraction tickNext = nextChordRest->tick() - baseTick;
-    if (tuplet) {
-        // for tuplets with odd ratios, apply ratio
-        // for tuplets with even ratios, use actual beat
-        Fraction ratio = tuplet->ratio();
-        if (ratio.numerator() & 1) {
-            tickNext *= ratio;
-        }
-    }
-
-    static const int BEAM_TUPLET_TOLERANCE = 6;
-    int tickLargeSize  = chord->ticks().ticks() * 2;
-    int remainder = tickNext.ticks() % tickLargeSize;
-    if (remainder <= BEAM_TUPLET_TOLERANCE || (tickLargeSize - remainder) <= BEAM_TUPLET_TOLERANCE) {
-        return true;
-    }
-
-    // default case
-    return false;
-}
-
-void Beam::createBeamletSegment(ChordRest* cr, bool isBefore, int level)
-{
-    const double startX = _layoutInfo.chordBeamAnchorX(cr,
-                                                       isBefore ? BeamTremoloLayout::ChordBeamAnchorType::End : BeamTremoloLayout::ChordBeamAnchorType::Start);
-
-    const double beamletLength = score()->styleMM(Sid::beamMinLen).val()
-                                 * cr->mag();
-
-    const double endX = startX + (isBefore ? -beamletLength : beamletLength);
-
-    double startY = _slope * (startX - _startAnchor.x()) + _startAnchor.y() - pagePos().y();
-    double endY = _slope * (endX - startX) + startY;
-
-    // how many beams past level 0 (i.e. beams on the other side of level 0 for this subgroup)
-    int extraBeamAdjust = 0;
-
-    // avoid adjusting for beams on opposite side of level 0
-    for (const BeamSegment* beam : _beamSegments) {
-        if (beam->level == 0 || beam->line.x2() < startX || beam->line.x1() > endX) {
-            continue;
-        }
-
-        if (cr->up() == beam->above) {
-            extraBeamAdjust++;
-        }
-    }
-
-    const int upValue = cr->up() ? -1 : 1;
-    const double verticalOffset = _beamDist * (level - extraBeamAdjust) * upValue;
-
-    if (RealIsEqual(_grow1, _grow2)) {
-        startY -= verticalOffset * _grow1;
-        endY -= verticalOffset * _grow1;
-    } else {
-        // Feathered beams
-        double startProportionAlongX = (startX - _startAnchor.x()) / (_endAnchor.x() - _startAnchor.x());
-        double endProportionAlongX = (endX - _startAnchor.x()) / (_endAnchor.x() - _startAnchor.x());
-
-        double grow1 = startProportionAlongX * (_grow2 - _grow1) + _grow1;
-        double grow2 = endProportionAlongX * (_grow2 - _grow1) + _grow1;
-
-        startY -= verticalOffset * grow1;
-        endY -= verticalOffset * grow2;
-    }
-
-    BeamSegment* b = new BeamSegment(this);
-    b->above = !cr->up();
-    b->level = level;
-    b->line = LineF(startX, startY, endX, endY);
-    b->isBeamlet = true;
-    b->isBefore = isBefore;
-    cr->setBeamlet(b);
-    _beamSegments.push_back(b);
-}
-
 void Beam::calcBeamBreaks(const ChordRest* chord, const ChordRest* prevChord, int level, bool& isBroken32, bool& isBroken64) const
 {
     BeamMode beamMode = chord->beamMode();
@@ -866,394 +382,6 @@ void Beam::calcBeamBreaks(const ChordRest* chord, const ChordRest* prevChord, in
             }
         }
     }
-}
-
-void Beam::createBeamSegments(const std::vector<ChordRest*>& chordRests)
-{
-    DeleteAll(_beamSegments);
-    _beamSegments.clear();
-
-    bool levelHasBeam = false;
-    int level = 0;
-    do {
-        levelHasBeam = false;
-        ChordRest* startCr = nullptr;
-        ChordRest* endCr = nullptr;
-        bool breakBeam = false;
-        bool previousBreak32 = false;
-        bool previousBreak64 = false;
-
-        for (size_t i = 0; i < chordRests.size(); i++) {
-            ChordRest* chordRest = chordRests[i];
-            ChordRest* prevChordRest = i < 1 ? nullptr : chordRests[i - 1];
-
-            if (level < chordRest->beams()) {
-                levelHasBeam = true;
-            }
-            bool isBroken32 = false;
-            bool isBroken64 = false;
-            // updates isBroken32 and isBroken64
-            calcBeamBreaks(chordRest, prevChordRest, level, isBroken32, isBroken64);
-            breakBeam = isBroken32 || isBroken64;
-
-            if (level < chordRest->beams() && !breakBeam) {
-                endCr = chordRest;
-                if (!startCr) {
-                    startCr = chordRest;
-                }
-            } else {
-                if (startCr && endCr) {
-                    if (startCr == endCr && startCr->isChord()) {
-                        bool isBeamletBefore = calcIsBeamletBefore(toChord(
-                                                                       startCr), static_cast<int>(i) - 1, level, previousBreak32,
-                                                                   previousBreak64);
-                        createBeamletSegment(toChord(startCr), isBeamletBefore, level);
-                    } else {
-                        createBeamSegment(startCr, endCr, level);
-                    }
-                }
-                bool setCr = chordRest && chordRest->isChord() && breakBeam && level < chordRest->beams();
-                startCr = setCr ? chordRest : nullptr;
-                endCr = setCr ? chordRest : nullptr;
-            }
-            previousBreak32 = isBroken32;
-            previousBreak64 = isBroken64;
-        }
-
-        // if the beam ends on the last chord
-        if (startCr && (endCr || breakBeam)) {
-            if ((startCr == endCr || !endCr) && startCr->isChord()) {
-                // this chord is either the last chord, or the first (followed by a rest)
-                bool isBefore = !(startCr == chordRests.front());
-                createBeamletSegment(toChord(startCr), isBefore, level);
-            } else {
-                createBeamSegment(startCr, endCr, level);
-            }
-        }
-        level++;
-    } while (levelHasBeam);
-}
-
-//---------------------------------------------------------
-//   layout2
-//---------------------------------------------------------
-
-void Beam::layout2(const std::vector<ChordRest*>& chordRests, SpannerSegmentType, int frag)
-{
-    _layoutInfo = BeamTremoloLayout(this);
-    Chord* startChord = nullptr;
-    Chord* endChord = nullptr;
-    if (chordRests.empty()) {
-        return;
-    }
-    for (auto chordRest : chordRests) {
-        if (chordRest->isChord()) {
-            if (!startChord) {
-                startChord = toChord(chordRest);
-                endChord = startChord;
-            } else {
-                endChord = toChord(chordRest);
-            }
-            toChord(chordRest)->layoutStem();
-        }
-    }
-    if (!startChord) {
-        // we were passed a vector of only rests. we don't support beams across only rests
-        // this beam will be deleted in LayoutBeams
-        return;
-    }
-
-    _beamSpacing = score()->styleB(Sid::useWideBeams) ? 4 : 3;
-    _beamDist = (_beamSpacing / 4.0) * spatium() * mag();
-    _beamWidth = point(score()->styleS(Sid::beamWidth)) * mag();
-
-    _startAnchor = _layoutInfo.chordBeamAnchor(startChord, BeamTremoloLayout::ChordBeamAnchorType::Start);
-    _endAnchor = _layoutInfo.chordBeamAnchor(endChord, BeamTremoloLayout::ChordBeamAnchorType::End);
-
-    if (_isGrace) {
-        _beamDist *= score()->styleD(Sid::graceNoteMag);
-        _beamWidth *= score()->styleD(Sid::graceNoteMag);
-    }
-
-    int fragmentIndex = (_direction == DirectionV::AUTO || _direction == DirectionV::DOWN) ? 0 : 1;
-    if (_userModified[fragmentIndex]) {
-        _layoutInfo = BeamTremoloLayout(this);
-        double startY = fragments[frag]->py1[fragmentIndex];
-        double endY = fragments[frag]->py2[fragmentIndex];
-        if (score()->styleB(Sid::snapCustomBeamsToGrid)) {
-            const double quarterSpace = spatium() / 4;
-            startY = round(startY / quarterSpace) * quarterSpace;
-            endY = round(endY / quarterSpace) * quarterSpace;
-        }
-        startY += pagePos().y();
-        endY += pagePos().y();
-        _startAnchor.setY(startY);
-        _endAnchor.setY(endY);
-        _layoutInfo.setAnchors(_startAnchor, _endAnchor);
-        _slope = (_endAnchor.y() - _startAnchor.y()) / (_endAnchor.x() - _startAnchor.x());
-        createBeamSegments(chordRests);
-        setTremAnchors();
-        return;
-    }
-
-    // anchor represents the middle of the beam, not the tip of the stem
-    // location depends on _isBesideTabStaff
-
-    if (!_isBesideTabStaff) {
-        _layoutInfo = BeamTremoloLayout(this);
-        _layoutInfo.calculateAnchors(chordRests, _notes);
-        _startAnchor = _layoutInfo.startAnchor();
-        _endAnchor = _layoutInfo.endAnchor();
-        _slope = (_endAnchor.y() - _startAnchor.y()) / (_endAnchor.x() - _startAnchor.x());
-        _beamDist = _layoutInfo.beamDist();
-    } else {
-        _slope = 0;
-        Chord* startChord = nullptr;
-        for (ChordRest* cr : chordRests) {
-            if (cr->isChord()) {
-                startChord = toChord(cr);
-                break;
-            }
-        }
-        _layoutInfo = BeamTremoloLayout(this);
-        double x1 = _layoutInfo.chordBeamAnchorX(chordRests.front(), BeamTremoloLayout::ChordBeamAnchorType::Start);
-        double x2 = _layoutInfo.chordBeamAnchorX(chordRests.back(), BeamTremoloLayout::ChordBeamAnchorType::End);
-        double y = _layoutInfo.chordBeamAnchorY(startChord);
-        _startAnchor = PointF(x1, y);
-        _endAnchor = PointF(x2, y);
-        _layoutInfo.setAnchors(_startAnchor, _endAnchor);
-        _beamWidth = _layoutInfo.beamWidth();
-    }
-
-    fragments[frag]->py1[fragmentIndex] = _startAnchor.y() - pagePos().y();
-    fragments[frag]->py2[fragmentIndex] = _endAnchor.y() - pagePos().y();
-
-    createBeamSegments(chordRests);
-    setTremAnchors();
-}
-
-bool Beam::layout2Cross(const std::vector<ChordRest*>& chordRests, int frag)
-{
-    int fragmentIndex = (_direction == DirectionV::AUTO || _direction == DirectionV::DOWN) ? 0 : 1;
-    ChordRest* startCr = _elements.front();
-    ChordRest* endCr = _elements.back();
-
-    const double quarterSpace = spatium() / 4;
-    // imagine a line of beamed notes all in a row on the same staff. the first and last of those
-    // are the 'outside' notes, and the slant of the beam is going to be affected by the 'middle' notes
-    // between them.
-    // we have to keep track of this for both staves.
-    Chord* topFirst = nullptr;
-    Chord* topLast = nullptr;
-    Chord* bottomFirst = nullptr;
-    Chord* bottomLast = nullptr;
-    int maxMiddleTopLine = std::numeric_limits<int>::min(); // lowest note in the top staff
-    int minMiddleBottomLine = std::numeric_limits<int>::max(); // highest note in the bottom staff
-    int prevTopLine = maxMiddleTopLine; // previous note's line position (top)
-    int prevBottomLine = minMiddleBottomLine; // previous note's line position (bottom)
-    // if the immediate neighbor of one of the two 'outside' notes on either the top or bottom
-    // are the same as that outside note, we need to record it so that we can add a 1/4 space slant.
-    bool secondTopIsSame = false;
-    bool secondBottomIsSame = false;
-    bool penultimateTopIsSame = false;
-    bool penultimateBottomIsSame = false;
-    double maxY = std::numeric_limits<double>::max();
-    double minY = std::numeric_limits<double>::min();
-    int otherStaff = 0;
-    // recompute _minMove and _maxMove as they may have shifted since last layout
-    _minMove = std::numeric_limits<int>::max();
-    _maxMove = std::numeric_limits<int>::min();
-    for (ChordRest* c : chordRests) {
-        IF_ASSERT_FAILED(c) {
-            continue;
-        }
-        int staffMove = c->staffMove();
-        _minMove = std::min(_minMove, staffMove);
-        _maxMove = std::max(_maxMove, staffMove);
-
-        if (staffMove != 0) {
-            otherStaff = staffMove;
-        }
-    }
-    if (otherStaff == 0 || _minMove == _maxMove) {
-        return false;
-    }
-    // Find the notes on the top and bottom of staves
-    //
-    bool checkNextTop = false;
-    bool checkNextBottom = false;
-    for (ChordRest* cr : chordRests) {
-        if (!cr->isChord()) {
-            continue;
-        }
-        Chord* c = toChord(cr);
-        if ((c->staffMove() == otherStaff && otherStaff > 0) || (c->staffMove() != otherStaff && otherStaff < 0)) {
-            // this chord is on the bottom staff
-            if (penultimateBottomIsSame) {
-                // the chord we took as the penultimate bottom note wasn't.
-                // so treat it properly as a middle note
-                minMiddleBottomLine = std::min(minMiddleBottomLine, prevBottomLine);
-                penultimateBottomIsSame = false;
-            }
-            checkNextTop = false; // we are no longer looking for the second note in the top
-                                  // staff being the same as the first--this note is on the bottom.
-            if (!bottomFirst) {
-                bottomFirst = c;
-                checkNextBottom = true; // this was the first bottom note, so check for second next time
-            } else {
-                penultimateBottomIsSame = prevBottomLine == c->line();
-                if (!penultimateBottomIsSame) {
-                    minMiddleBottomLine = std::min(minMiddleBottomLine, prevBottomLine);
-                }
-                if (checkNextBottom) {
-                    // this is the second bottom note, so we should see if this one is same line as first
-                    secondBottomIsSame = c->line() == bottomFirst->line();
-                    checkNextBottom = false;
-                } else {
-                    prevBottomLine = c->line();
-                }
-                bottomLast = c;
-            }
-            maxY = std::min(maxY, _layoutInfo.chordBeamAnchorY(toChord(c)));
-        } else {
-            // this chord is on the top staff
-            if (penultimateTopIsSame) {
-                // the chord we took as the penultimate top note wasn't.
-                // so treat it properly as a middle note
-                maxMiddleTopLine = std::max(maxMiddleTopLine, prevTopLine);
-                penultimateTopIsSame = false;
-            }
-            checkNextBottom = false; // no longer looking for a bottom second note since this is on top
-            if (!topFirst) {
-                topFirst = c;
-                checkNextTop = true;
-            } else {
-                penultimateTopIsSame = prevTopLine == c->line();
-                if (!penultimateTopIsSame) {
-                    maxMiddleTopLine = std::max(maxMiddleTopLine, prevTopLine);
-                }
-                if (checkNextTop) {
-                    secondTopIsSame = c->line() == topFirst->line();
-                    checkNextTop = false;
-                } else {
-                    prevTopLine = c->line();
-                }
-                topLast = c;
-            }
-            minY = std::max(minY, _layoutInfo.chordBeamAnchorY(toChord(c)));
-        }
-    }
-    _startAnchor.ry() = (maxY + minY) / 2;
-    _endAnchor.ry() = (maxY + minY) / 2;
-    _slope = 0;
-
-    if (!noSlope()) {
-        int topFirstLine = topFirst ? topFirst->downNote()->line() : 0;
-        int topLastLine = topLast ? topLast->downNote()->line() : 0;
-        int bottomFirstLine = bottomFirst ? bottomFirst->upNote()->line() : 0;
-        int bottomLastLine = bottomLast ? bottomLast->upNote()->line() : 0;
-        bool constrainTopToQuarter = false;
-        bool constrainBottomToQuarter = false;
-        if ((topFirstLine > topLastLine && secondTopIsSame)
-            || (topFirstLine < topLastLine && penultimateTopIsSame)) {
-            constrainTopToQuarter = true;
-        }
-        if ((bottomFirstLine < bottomLastLine && secondBottomIsSame)
-            || (bottomFirstLine > bottomLastLine && penultimateBottomIsSame)) {
-            constrainBottomToQuarter = true;
-        }
-        if (!topLast && !bottomLast && topFirst && bottomFirst) {
-            // if there are only two notes, one on each staff, special case
-            // take max slope into account
-            double yFirst, yLast;
-            if (topFirst->tick() < bottomFirst->tick()) {
-                yFirst = topFirst->stemPos().y();
-                yLast = bottomFirst->stemPos().y();
-            } else {
-                yFirst = bottomFirst->stemPos().y();
-                yLast = topFirst->stemPos().y();
-            }
-            int desiredSlant = round((yFirst - yLast) / spatium());
-            int slant = std::min(std::abs(desiredSlant), _layoutInfo.getMaxSlope());
-            slant *= (desiredSlant < 0) ? -quarterSpace : quarterSpace;
-            _startAnchor.ry() += (slant / 2);
-            _endAnchor.ry() -= (slant / 2);
-        } else if (!topLast || !bottomLast) {
-            // otherwise, if there is only one note on one of the staves, use slope from other staff
-            int startNote = 0;
-            int endNote = 0;
-            bool forceHoriz = false;
-            if (!topLast) {
-                startNote = bottomFirstLine;
-                endNote = bottomLastLine;
-                if (minMiddleBottomLine <= std::min(startNote, endNote)) {
-                    // there is a note closer to the beam than the start and end notes
-                    // we force horizontal beam here.
-                    forceHoriz = true;
-                }
-            } else if (!bottomLast) {
-                startNote = topFirstLine;
-                endNote = topLastLine;
-                if (maxMiddleTopLine >= std::max(startNote, endNote)) {
-                    // same as above, for the top staff
-                    // force horizontal.
-                    forceHoriz = true;
-                }
-            }
-
-            if (!forceHoriz) {
-                int slant = startNote - endNote;
-                slant = std::min(std::abs(slant), _layoutInfo.getMaxSlope());
-                if ((!bottomLast && constrainTopToQuarter) || (!topLast && constrainBottomToQuarter)) {
-                    slant = 1;
-                }
-                double slope = slant * (startNote > endNote ? quarterSpace : -quarterSpace);
-                _startAnchor.ry() += (slope / 2);
-                _endAnchor.ry() -= (slope / 2);
-            } // otherwise, do nothing, beam is already horizontal.
-        } else {
-            // otherwise, there are at least two notes on each staff
-            // (that is, topLast and bottomLast are both set)
-            bool forceHoriz = false;
-            if (topFirstLine == topLastLine || bottomFirstLine == bottomLastLine) {
-                // if outside notes on top or bottom staff are on the same staff line, slope = 0
-                // no further adjustment needed, the beam is already well-placed and horizontal
-                forceHoriz = true;
-            }
-            // otherwise, we have to compare the slopes from the top staff and bottom staff.
-            int topSlant = topFirstLine - topLastLine;
-            if (constrainTopToQuarter && topSlant != 0) {
-                topSlant = topFirstLine < topLastLine ? -1 : 1;
-            }
-            int bottomSlant = bottomFirstLine - bottomLastLine;
-            if (constrainBottomToQuarter && bottomSlant != 0) {
-                bottomSlant = bottomFirstLine < bottomLastLine ? -1 : 1;
-            }
-            if ((maxMiddleTopLine >= std::max(topFirstLine, topLastLine)
-                 || (minMiddleBottomLine <= std::min(bottomFirstLine, bottomLastLine)))) {
-                forceHoriz = true;
-            }
-            if (topSlant == 0 || bottomSlant == 0 || forceHoriz) {
-                // if one of the slants is 0, the whole slant is zero
-            } else if ((topSlant < 0 && bottomSlant < 0) || (topSlant > 0 && bottomSlant > 0)) {
-                int slant = (abs(topSlant) < abs(bottomSlant)) ? topSlant : bottomSlant;
-                slant = std::min(std::abs(slant), _layoutInfo.getMaxSlope());
-                double slope = slant * ((topSlant < 0) ? -quarterSpace : quarterSpace);
-                _startAnchor.ry() += (slope / 2);
-                _endAnchor.ry() -= (slope / 2);
-            } else {
-                // if the two slopes are in opposite directions, flat!
-                // nothing needs to be done, the beam is already horizontal and placed nicely
-            }
-        }
-        _startAnchor.setX(_layoutInfo.chordBeamAnchorX(startCr, BeamTremoloLayout::ChordBeamAnchorType::Start));
-        _endAnchor.setX(_layoutInfo.chordBeamAnchorX(endCr, BeamTremoloLayout::ChordBeamAnchorType::End));
-        _slope = (_endAnchor.y() - _startAnchor.y()) / (_endAnchor.x() - _startAnchor.x());
-    }
-    fragments[frag]->py1[fragmentIndex] = _startAnchor.y() - pagePos().y();
-    fragments[frag]->py2[fragmentIndex] = _endAnchor.y() - pagePos().y();
-    createBeamSegments(chordRests);
-    return true;
 }
 
 //---------------------------------------------------------
