+++ src/engraving/libmscore/spanner.h
@@ -26,18 +26,11 @@
 #include <deque>
 
 #include "draw/types/color.h"
+#include "types/types.h"
 #include "engravingitem.h"
 
 namespace mu::engraving {
 class Spanner;
-
-//---------------------------------------------------------
-//   SpannerSegmentType
-//---------------------------------------------------------
-
-enum class SpannerSegmentType {
-    SINGLE, BEGIN, MIDDLE, END
-};
 
 //---------------------------------------------------------
 //   @@ SpannerSegment
